+++ app/components/FireBaseMessaging.js
@@ -1,5 +1,5 @@
 import React, { Component, Fragment } from 'react';
-import { Alert, Platform, AppState, AppStateStatus, Linking } from 'react-native';
+import { Alert, Platform, AppState, AppStateStatus, Linking, NativeModules } from 'react-native';
 import PropTypes from 'prop-types';
 import * as firebase from 'react-native-firebase';
 import { connect } from 'react-redux';
@@ -38,6 +38,9 @@ class FireBaseMessaging extends Component {
    * @returns {void}
    */
   async componentDidMount() {
+    if (AppState.currentState === 'background' && isAndroid) {
+      NativeModules.DevSettings.reload();
+    }
     this.listeners = [
       fNotifications.onNotification(this.onNotification),
       fNotifications.onNotificationDisplayed(this.onNotificationDisplayed),
@@ -66,6 +69,7 @@ class FireBaseMessaging extends Component {
 
     if (event) {
       this.openActivityByEventId(event);
+      // if (isAndroid) NativeModules.DevSettings.reload();
     }
   }